+++ obj/Release/net10.0/SealedFates.AssemblyInfo.cs
@@ -12,7 +12,7 @@ using System;
 using System.Reflection;
 
 [assembly: System.Reflection.AssemblyCompanyAttribute("SealedFates")]
-[assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
+[assembly: System.Reflection.AssemblyConfigurationAttribute("Release")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
 [assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+f2bc3b0cdf1730190f16cce550141588d3de7995")]
 [assembly: System.Reflection.AssemblyProductAttribute("SealedFates")]